split visualization.py
Use separate file for emitting dot and emitting asciidag.
